以 2W 元素为基准测试
> 优化前 `1175 ms`

- [x] 优化 `mergeOptions` 性能  (`1050 ms`)
- [x] 撤销特殊情况下的样式拷贝优化，并不能提升性能 (`880 ms`) https://github.com/antvis/G6/pull/6651
- [x] 移除 `effect` 优化逻辑 (`860 ms`)
- [x] 关闭用例主题 (`814 ms`)

整体时长缩短 `361 ms` (`30%`)